task33

+++ task3/Program.cs
@@ -1,12 +1,12 @@
-﻿double[] array = new double[10];
+﻿int[] array = new int[10];
 for (int i = 0; i < array.Length; i++)
 {
 array[i] = new Random().Next(100,1000);
 Console.WriteLine(array[i]);
 }
 
-double max = array [0];
-double min = array [0];
+int max = array [0];
+int min = array [0];
 
 for (int i = 0; i < array.Length; i++)
 {
